+++ app/src/main/java/com/episode6/meetingminder/store/AppStoreReducer.kt
@@ -1,0 +1,14 @@
+package com.episode6.meetingminder.store
+
+import com.episode6.redux.Action
+
+internal fun AppState.reduce(action: Action): AppState = when (action) {
+    is UpdateStateAction -> reduceUpdateStateAction(action)
+    else -> this
+}
+
+private fun AppState.reduceUpdateStateAction(action: UpdateStateAction): AppState = when (action) {
+    is SetSettledDate -> copy(settledDate = action.date)
+    is ShowMessage -> copy(transientMessage = action.message)
+    is ClearMessage -> if (transientMessage?.id == action.id) copy(transientMessage = null) else this
+}