+++ app/src/main/java/com/episode6/meetingminder/ui/theme/Theme.kt
@@ -52,6 +52,7 @@ fun MeetingMinderTheme(
 ) {
     MaterialTheme(
         colorScheme = if (darkTheme) DarkColors else LightColors,
+        typography = MeetingMinderTypography,
         content = content,
     )
 }